🐛Force usage of chokidar 3
Motivation
----------

fixes https://github.com/mswjs/msw/issues/188

The current version of webpack that we're using has [a bug][1] that
causes the tests to crashes on MacOS with some regularity. That is
because it depends on chokidar 2 which uses the `fsevents` api which
has a [crictical bug][2] on MacOS.

Approach
--------

This manually forces all modules to use chokidar 3 which does not use
the `fsevents` module at all. While this does trigger a semver
violation because webpack 4 depends on chokidar 2, the API is
apparently similar enough that all the tests pass. It does however
give rise to a warning in the console upon installation of depedencies
via a `yarn`.

This resolution can be removed once an upgrade is made to webpack 5 or
to a different depedency bundler like [parcel][3] or [rollup][4].

Learning
--------

* [Selective dependency resolutions ][5]

[1]: https://github.com/webpack/webpack/issues/10037
[2]: https://github.com/fsevents/fsevents/issues/314
[3]: http://parceljs.org
[4]: https://rollupjs.org
[5]: https://classic.yarnpkg.com/en/docs/selective-version-resolutions/