content-type header parsing: be more forgiving
The space after the semicolon delimiter in HTTP header content is optional according to the standard, and a number of implementations will not add such whitespace.

I ran into this problem when mocking form data using the form-data lib (https://www.npmjs.com/package/form-data), which does not add the optional space when serializing headers.