refactor: graphql imported by import instead of require
removes the previous attempt at making graphql optional by lazily requiring it instead of importing
it.  This didn't have the intended affect of helping to avoid parsing in many environments, and
might be confusing and/or cause friction with a future migration to using native modules and
providing an export that can be treeshaken in all environments.

re #1371