fix: properly handle concurrent requests while stopping the service worker
If you stop the service worker while requests are in-flight then then are not properly handled so
the requests never resolve. This change makes sure that we properly wait for a ack from the service
worker got the stop request and that any in-flight requests are resolved as passthough.

BREAKING CHANGE: The stop API goes from being sync to async.

fix #2323